Specify host for client login during init to ignore MYSQL_HOST value
When the server is started up during initialization, it will only allow access
through the localhost socket interface. The entrypoint script logs in and changes
the root user to allow remote access, but if the MYSQL_HOST environment variable is
defined it will be used by the client, causing the script's login to fail if it
is set to anything other than localhost.

Adding a -hlocalhost to the client login during init makes it ignore the environment variable.