Hardcode socket path during init
During container init, we start both server and client temporarily,
for user admin etc. For the server we pass on any command line
arguments specified by the user, but for the client we don't. This
can cause an issue if the user specifies a different socket path,
since the client won't find it, causing init failure. Since the
socket location doesn't really matter for these temporary runs, we
hardcode it to /var/run/mysqld/mysqld.sock