Fetch client --socket value from config instead of hardcoding it
Hardcoding the socket path for the init run broke initdb.d/ scripts that
expected a different location (because 5.5 has a different default or the user
changed the config).
Added a function for fetching it with mysqld --verbose --help as we do for
datadir, and use that for the initialization runs of the server and client.