Touch $DATADIR/docker-entrypoint-complete after loading complete
When the mysql container is autoloading data from
/docker-entrypoint-initdb.d/ it's difficult to know when the load has
completed. This is particularly true in a kubernetes
liveliness/readiness check where we don't know if the container is ready
to serve yet.

After the load has finished we touch a file that we can test for the
presence of before responding the container is ready.