Allow to configure DATADIR via env variable
This is useful to do when mounting volumes. In some cases, e.g.
with named volumes, it's not possible to mount sub-folder https://github.com/moby/moby/issues/32582. This limitation prevents reuse of same volume. For example, If both mongodb and mysql are configured to use shared volume as its DB folder, mysql will refuse to start, since its target DB folder is not empty ([ERROR] --initialize specified but the data directory has files in it. Aborting.). This patch allows to solve this by mounting shared volume to different directory than /var/lib/mysql and define $MYSQL_DATADIR to subfolder of this shared volume.